Added constants and classes needed for the game
I added GAME_WIDTH and GAME_HEIGHT to measure the width and height of the game. SPEED determines the quickness of the game. SPACE_SIZE determines how large the items like the food and snake are in the game. BODY_PARTS is the amount of body parts that the snake starts with. SNAKE_COLOR, FOOD_COLOR, and BACKROUND_COLOR are pretty self explanatory.